// blueprints/orion-platform-core/src/types/core.ts
export interface Project {
  id: string;
  tenantId: string;
  name: string;
  slug: string;
  description: string | null;
  status: ProjectStatus;
  createdAt: Date;
  updatedAt: Date;
}

export type ProjectStatus = 'active' | 'archived' | 'deleted';

export type CreateProjectInput = {
  name: string;
  slug: string;
  description?: string;
};

export type UpdateProjectInput = Partial<{
  name: string;
  slug: string;
  description: string | null;
  status: ProjectStatus;
}>;